QCMaterial

QCMaterial.QulacsVariationalQuantumCircuit

Base.copy

QCMaterial.apply_qubit_op!

QCMaterial.check_py_type

QCMaterial.convert_openfermion_op

QCMaterial.divide_real_imag

QCMaterial.gen_p_t2

QCMaterial.gen_t1

QCMaterial.generate_numerical_grad

QCMaterial.get_thetas

QCMaterial.get_transition_amplitude_with_obs

QCMaterial.imag_time_evolve

QCMaterial.mk_scipy_minimize

QCMaterial.num_theta

QCMaterial.numerical_grad

QCMaterial.overlap

QCMaterial.parse_pauli_str

QCMaterial.solve_gs

QCMaterial.tikhonov

QCMaterial.transition

QCMaterial.uccgsd

QCMaterial.update_circuit_param!

QCMaterial.update_circuit_param!

QCMaterial.update_quantum_state!

QCMaterial.update_quantum_state!

QCMaterial.update_quantum_state!

QCMaterial.QulacsVariationalQuantumCircuit — Type

Wrap a QulacsParametricQuantumCircuit object, which will not be copied.

Base.copy — Method

Copy a QulacsVariationalQuantumCircuit object. This makes a copy of the underlying QulacsParametricQuantumCircuit object.

QCMaterial.apply_qubit_op! — Method

Apply a qubit operator op to |stateket> and fit the result with circuit * |statebra>. The circuit object will be updated on exit. The squared norm of op * |stateket> will be returned. state0bra will not be modified.

QCMaterial.check_py_type — Method

Check if the Python type of a given PyObject matches the expected one

QCMaterial.convert_openfermion_op — Method

Convertopenfermionop

Args: nqubit (:class:int) openfermionop (:class:openfermion.ops.QubitOperator) Returns: :class:qulacs.Observable

QCMaterial.divide_real_imag — Method

Divide a qubit operator into the hermite and antihermite parts.

QCMaterial.gen_p_t2 — Method

Generate pair dobule excitations

QCMaterial.gen_t1 — Method

Generate single excitations

QCMaterial.generate_numerical_grad — Method

Generates parallelized numerical grad_cost

QCMaterial.get_thetas — Method

Return a copy of variational parameters

QCMaterial.get_transition_amplitude_with_obs — Method

Compute <statebra| circuit^+ obs |stateket>, where obs is a hermite observable.

QCMaterial.imag_time_evolve — Function

Perform imaginary-time evolution.

ham_op: Hamiltonian vc: Variational circuit. The current value of variational parameters are used as the initial value of the imaginary-time evolution. state0: The initial state to which the Variational circuit is applied to taus: list of imaginary times in ascending order The first element must be 0.0. return: list of variational parameters at the given imaginary times.

QCMaterial.mk_scipy_minimize — Function

Make a wrapped scipy minimizer

QCMaterial.num_theta — Method

Return the number of independent variational parameters

QCMaterial.numerical_grad — Method

Compute partial derivative of a given function at a point x

QCMaterial.overlap — Method

Compute <phi (thetabra) | phi(thetaket)>

QCMaterial.parse_pauli_str — Method

Parse a tuple representing a Pauli string When x is ((1, "X"), (5, "Y")), returns [1, 5], [PauliID.X, PauliID.Y]

QCMaterial.solve_gs — Method

Compute ground state

QCMaterial.tikhonov — Function

y = Ax tikhov regularization: minimize ||Ax-b||^2 + λ||x^2||

QCMaterial.transition — Method

Compute <phi (thetabra) |H| phi(thetaket)>

QCMaterial.uccgsd — Method

Returns UCCGSD circuit.

QCMaterial.update_circuit_param! — Method

Update circuit parameters

thetas wil be copied.

QCMaterial.update_circuit_param! — Method

Update the values of the independent variational parameters

QCMaterial.update_quantum_state! — Method

Update a state using a circuit

QCMaterial.update_quantum_state! — Method

Update a state using a circuit

QCMaterial.update_quantum_state! — Method

Update a state using a circuit